+++ R/reliability_data.R
@@ -26,7 +26,6 @@
 #' }
 #' If \code{.keep_all = TRUE}, the remaining columns of \code{data} are also preserved.
 #'
-#'
 #' @examples
 #' # Example 1 -  Based on an existing data.frame/tibble and column names:
 #' data <- reliability_data(
@@ -51,13 +50,8 @@
 #' )
 #'
 #' # Example 4 - Based on vectors:
-#' cycles   <- c(300, 300, 300, 300, 300, 291, 274, 271, 269, 257, 256, 227, 226,
-#'               224, 213, 211, 205, 203, 197, 196, 190, 189, 188, 187, 184, 180,
-#'               180, 177, 176, 173, 172, 171, 170, 170, 169, 168, 168, 162, 159,
-#'               159, 159, 159, 152, 152, 149, 149, 144, 143, 141, 141, 140, 139,
-#'               139, 136, 135, 133, 131, 129, 123, 121, 121, 118, 117, 117, 114,
-#'               112, 108, 104, 99, 99, 96, 94)
-#' state <- c(rep(0, 5), rep(1, 67))
+#' cycles <- alloy$cycles
+#' state <- alloy$status
 #' id <- "XXXXXX"
 #'
 #' data_4 <- reliability_data(
@@ -127,7 +121,7 @@ print.wt_reliability_data <- function(x, ...) {
     cat("Reliability Data:\n")
   } else {
     cat(
-      "Reliability data with characteristic '",
+      "Reliability Data with characteristic '",
       attr(x, "characteristic"),
       "':\n",
       sep = ""